+++ packages/nano-ui/test/integration/Main.hs
@@ -35,6 +35,7 @@ import Cases.PointerOwnership qualified
 import Cases.Tooltip qualified
 import Cases.CustomWidget qualified
 import Cases.ShrinkDamage qualified
+import Cases.WidgetIdIndex qualified
 import NanoUI.Testing.Runner (runTests)
 
 main :: IO ()
@@ -75,4 +76,5 @@ main =
     , Cases.Tooltip.tests
     , Cases.CustomWidget.tests
     , Cases.ShrinkDamage.tests
+    , Cases.WidgetIdIndex.tests
     ]